Taiga #70 Implement Circular Collision
Currently implemented is an early circle collision algorithm that detects n
amount of objects in the collidable_list table in src/client_game.lua

A demo is set up using two wall objects at the bridge to demonstrate.
There is room to improve the system and way of identifying which objects
should be collidable or not as well as how a bullet or skill uses the
collision algorithm as well.